remove dotty.dokka.model.api package
